написал тестер небольшой для инта, работаю пока с int

+++ ft_printf.h
@@ -41,7 +41,7 @@ void			ft_putnbr_u(unsigned int n);
 int				ft_printf(char *argv1, ...);
 int				ft_isdigit(int c);
 int				ft_atoi(char *str);
-void			ft_putnbr(int n);
+void			ft_putnbr(long n);
 void			display_int(t_struct *box);
 size_t			ft_rank_count(long n, int base);
 void			ft_put_null_to_struct(t_struct *box);
